fixes volumes load

+++ entrypoint.sh
@@ -6,6 +6,8 @@ set -e
 
 FIRST_START_DONE="/etc/docker-openldap-first-start-done"
 
+[ ! -e /var/lib/openldap/data ] && mkdir -p /var/lib/openldap/data
+
 # container first start
 if [ ! -e "$FIRST_START_DONE" ]; then
 
@@ -21,26 +23,18 @@ if [ ! -e "$FIRST_START_DONE" ]; then
         exit 1
     fi
 
-  if [[ ! -d /var/lib/openldap/data ]]; then
-    mkdir -p /var/lib/openldap/data
-    cp /etc/openldap/DB_CONFIG.example /var/lib/openldap/data/DB_CONFIG
-    chown -R ldap:ldap /var/lib/openldap/data
-    chmod 700 /var/lib/openldap/data
-  fi
-
-  if [[ ! -d /etc/openldap/slapd.d ]]; then
+  if [ ! -d /etc/openldap/slapd.d ]; then
     mkdir /etc/openldap/slapd.d
   fi
 
-  chown -R ldap:ldap /var/lib/openldap/ /etc/openldap/slapd.d
+  chown -R ldap:ldap /etc/openldap/slapd.d
 
-  if [[ -n "$SLAPD_PASSWORD" ]]; then
+  if [ -n "$SLAPD_PASSWORD" ]; then
     password_hash=`slappasswd -s "${SLAPD_PASSWORD}"`
     sed_safe_password_hash=${password_hash//\//\\\/}
     sed -i "s|rootpw.*|rootpw        ${sed_safe_password_hash}|g" /etc/openldap/slapd.conf
   fi
 
-  if [[ -n "$SLAPD_DOMAIN" ]]; then
     SLAPD_ORGANIZATION="${SLAPD_ORGANIZATION:-${SLAPD_DOMAIN}}"
 
     dc_string=""
@@ -60,17 +54,19 @@ if [ ! -e "$FIRST_START_DONE" ]; then
     sed -i "s|o: Example|o: $SLAPD_ORGANIZATION|g" /etc/openldap/modules/base.ldif
     sed -i "s/^#BASE.*/BASE  ${base_string}/g" /etc/openldap/ldap.conf
 
-    slapd -u ldap -g ldap >/dev/null 2>&1
-
-    ldapadd -x -D "cn=admin,${base_string}" -w "${SLAPD_PASSWORD}" -f /etc/openldap/modules/base.ldif
-
-    killall slapd
-
-    # slapadd -n0 -l /etc/openldap/modules/base.ldif >/dev/null 2>&1
+  if [ -f /etc/conf.d/slapd ]; then
+    sed -i "s|#OPTS=.*|OPTS=\"-F /etc/openldap/slapd.d -h 'ldap:// ldapi://%2fvar%2frun%2fopenldap%2fslapd.sock'\"|g" /etc/conf.d/slapd
   fi
 
-  if [[ -f /etc/conf.d/slapd ]]; then
-    sed -i "s|#OPTS=.*|OPTS=\"-F /etc/openldap/slapd.d -h 'ldap:// ldapi://%2fvar%2frun%2fopenldap%2fslapd.sock'\"|g" /etc/conf.d/slapd
+    chown -R ldap:ldap /var/lib/openldap/data
+    chmod 700 /var/lib/openldap/data
+  if [ -z "$(ls -A /var/lib/openldap/data)" ]; then
+    echo "data directory is empty, init..."
+    cp /etc/openldap/DB_CONFIG.example /var/lib/openldap/data/DB_CONFIG
+
+    slapd -u ldap -g ldap >/dev/null 2>&1
+    ldapadd -x -D "cn=admin,${base_string}" -w "${SLAPD_PASSWORD}" -f /etc/openldap/modules/base.ldif
+    killall slapd
   fi
 
   touch $FIRST_START_DONE
